# Conflicts:
#	src/main/java/cn/edu/ldu/watercloud/entity/ChargeFormMonthly.java
#	src/main/java/cn/edu/ldu/watercloud/mapper/ProjectMapper.java
#	src/main/resources/application.yaml

+++ src/main/java/cn/edu/ldu/watercloud/entity/Watermeter.java
@@ -17,5 +17,5 @@ public class Watermeter {
     private double totalflow;
     private double lastflow;
     private double banalnce;
-    private double status;
+    private int status;
 }
